Move customer balance updates to an after_save callback
An `after_save` callback is wrapped in the same transaction
as the actual save operation. So any exceptions or nested transactions
that fail in an `after_save` callback, will cause the transfer transaction
to not be created in the database. Putting this logic in an `after_save` will
ensure that this step is never forgotten to be executed.